update(ui): split conversation
