--- app/(main)/diary.tsx
@@ -1,14 +1,0 @@
-import React, {Component} from 'react'
-import {Text, View} from 'react-native'
-
-export class diary extends Component {
-    render() {
-        return (
-            <View>
-                <Text> textInComponent </Text>
-            </View>
-        )
-    }
-}
-
-export default diary
